+++ mt/src/main/scala/org/platanios/symphony/mt/models/parameters/WordEmbeddingsPerLanguagePair.scala
@@ -140,6 +140,6 @@ object WordEmbeddingsPerLanguagePair {
       tgtLanguageId: Output
   ): Output = {
     srcLanguageId * (numLanguages - 1) + tgtLanguageId -
-        tf.less(srcLanguageId, tgtLanguageId).cast(INT64)
+        tf.less(srcLanguageId, tgtLanguageId).cast(INT32)
   }
 }
